Add files via upload
This program prompts the user to enter a key (shift value) and a message,
then uses the Caesar Cipher algorithm to encrypt and decrypt the message.
The encrypt and decrypt methods both take in a message and a key,
shift each character in the message by the key (or negative key for decryption), 
and return the resulting message as a string. Note that this program only works for letters (uppercase and lowercase),
and any other characters in the message will not be shifted.